python-fd: merge ldap plugin into a single file
