+++ core/src/plugins/filed/python/CMakeLists.txt
@@ -118,7 +118,6 @@ set(PYFILES
     pyfiles/BareosFdPluginLocalFilesBaseclass.py
     pyfiles/bareos-fd-local-fileset.py
     ldap/bareos-fd-ldap.py
-    ldap/BareosFdPluginLDAP.py
     mariabackup/BareosFdPluginMariabackup.py
     mariabackup/bareos-fd-mariabackup.py
     ovirt/bareos-fd-ovirt.py
