+++ debian/bareos-filedaemon-ldap-python-plugin.install.in
@@ -1,5 +1,4 @@
 @plugindir@/bareos-fd-ldap.py*
-@plugindir@/BareosFdPluginLDAP.py*
 @configtemplatedir@/bareos-dir.d/fileset/plugin-ldap.conf.example
 @configtemplatedir@/bareos-dir.d/job/backup-ldap.conf.example
 @configtemplatedir@/bareos-dir.d/job/restore-ldap.conf.example
